backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Redis
redis==5.2.1

# HTTP Client (Azure API calls)
httpx==0.28.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Settings
pydantic-settings==2.7.1

# Validation
email-validator==2.1.1

# Utilities
python-multipart==0.0.20
python-dotenv==1.0.1
